bank_Project/requirements.txt
flask
flasgger
flask-sqlalchemy
pymysql
python-dotenv
cryptography
